+++ src/Constants.cpp
@@ -45,12 +45,12 @@ namespace constants
     //        return(nMutRange);
     //    }
 
-    std::vector<unsigned int> Constants::setNMutRange(const unsigned int maxMut, const unsigned int L,
-                                                      const unsigned int q)
+    std::vector<unsigned long long> Constants::setNMutRange(const unsigned int maxMut, const unsigned int L,
+                                                            const unsigned int q)
     {
         // compute the number of possible sequence for 0..MAX_MUT mutations, the cumulative sum gives the id range for
         // each number of mutations larger range for adding errors (i.e. more mutated positions)
-        std::vector<unsigned int> nMutRange(maxMut * 2 + 1);
+        std::vector<unsigned long long> nMutRange(maxMut * 2 + 1);
         nMutRange[0] = 1;
         for (unsigned int i = 1; i <= maxMut * 2; ++i)
         {
